Removing unused components and starting to move some into utilities dir.
