+++ resources/assets/components/utilities/FormMessage/form-message.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .form-message {
   &.-error {
