+++ resources/assets/components/utilities/MediaUploader/index.js
@@ -4,7 +4,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-import { processFile } from '../../helpers';
+import { processFile } from '../../../helpers';
 
 import './media-uploader.scss';
